feat: replace AddMemberScreen with AddMembersScreen and implement member selection and search functionality
